requirements.txt

requests
disnake
asyncio
polygonsdk
webullsdk
psycopg2
